refactor: extract final forest identity policy
